+++ spec/osm/api_spec.rb
@@ -100,7 +100,7 @@ describe Osm::Api do
 
     it 'User is invalid' do
       allow($api).to receive(:valid_user?) { false }
-      expect { $api.require_valid_user! }.to raise_error(Osm::Api::UserInvalid, 'id: "2", secret: "USER-SECRET"')
+      expect { $api.require_valid_user! }.to raise_error(Osm::APIError::InvalidUser, 'id: "2", secret: "USER-SECRET"')
     end
   end # describe Requires a valid user
 
@@ -176,7 +176,7 @@ describe Osm::Api do
         response.content_type = 'giberish/nothing-meaningful'
         allow(response).to receive(:body).and_return('body')
         expect($http).to receive(:request).with($request).and_return(response)
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::Error, 'Unhandled content-type: giberish/nothing-meaningful')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::APIError::UnexpectedType, 'Got a: giberish/nothing-meaningful')
       end
     end
 
@@ -203,17 +203,17 @@ describe Osm::Api do
       it "Raises a connection error if the HTTP status code was not 'OK'" do
         response = Net::HTTPInternalServerError.new(1, 500, '')
         expect($http).to receive(:request).with($request).and_return(response)
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::ConnectionError, 'HTTP Status code was 500')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::APIError::ConnectionError, 'HTTP Status code was 500')
       end
 
       it "Raises a connection error if it can't connect to OSM" do
         expect($http).to receive(:request).with($request) { raise Errno::ENETUNREACH, 'Failed to open TCP connection to 2.127.245.223:80 (Network is unreachable - connect(2) for "2.127.245.223" port 80)' }
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::ConnectionError, 'Errno::ENETUNREACH: Network is unreachable - Failed to open TCP connection to 2.127.245.223:80 (Network is unreachable - connect(2) for "2.127.245.223" port 80)')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::APIError::ConnectionError, 'Errno::ENETUNREACH: Network is unreachable - Failed to open TCP connection to 2.127.245.223:80 (Network is unreachable - connect(2) for "2.127.245.223" port 80)')
       end
 
       it 'Raises a connection error if an SSL error occurs' do
         expect($http).to receive(:request).with($request) { raise OpenSSL::SSL::SSLError, 'hostname "2.127.245.223" does not match the server certificate' }
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::ConnectionError, 'OpenSSL::SSL::SSLError: hostname "2.127.245.223" does not match the server certificate')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::APIError::ConnectionError, 'OpenSSL::SSL::SSLError: hostname "2.127.245.223" does not match the server certificate')
       end
 
     end # Handles network errors
@@ -224,21 +224,21 @@ describe Osm::Api do
         expect($request).to receive(:set_form_data).with('apiid' => '1', 'token' => 'API-SECRET', 'userid' => '2', 'secret' => 'USER-SECRET').and_return(nil)
         allow($response).to receive(:body) { '{"error":"Error message"}' }
         expect($http).to receive(:request).with($request).and_return($response)
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::Error, 'Error message')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::OSMError, 'Error message')
       end
 
       it 'Raises an error if OSM returns an error (as a hash in a hash)' do
         expect($request).to receive(:set_form_data).with('apiid' => '1', 'token' => 'API-SECRET', 'userid' => '2', 'secret' => 'USER-SECRET').and_return(nil)
         allow($response).to receive(:body) { '{"error":{"message":"Error message"}}' }
         expect($http).to receive(:request).with($request).and_return($response)
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::Error, 'Error message')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::OSMError, 'Error message')
       end
 
       it 'Raises an error if OSM returns an error (as a plain string)' do
         expect($request).to receive(:set_form_data).with('apiid' => '1', 'token' => 'API-SECRET', 'userid' => '2', 'secret' => 'USER-SECRET').and_return(nil)
         allow($response).to receive(:body) { 'Error message' }
         expect($http).to receive(:request).with($request).and_return($response)
-        expect { $api.post_query('path/to/load') }.to raise_error(Osm::Error, 'Error message')
+        expect { $api.post_query('path/to/load') }.to raise_error(Osm::OSMError, 'Error message')
       end
 
     end # Handles OSM errors
@@ -263,19 +263,19 @@ describe Osm::Api do
     describe 'User has no roles in OSM' do
       it 'OSM returns false' do
         expect($api).to receive(:post_query).with('api.php?action=getUserRoles').twice { false }
-        expect { $api.get_user_roles! }.to raise_error(Osm::NoActiveRoles)
+        expect { $api.get_user_roles! }.to raise_error(Osm::OSMError::NoActiveRoles)
         expect($api.get_user_roles).to eq([])
       end
       it 'OSM causes an exception to be raised' do
-        expect($api).to receive(:post_query).with('api.php?action=getUserRoles').twice { fail Osm::Error, 'false' }
-        expect { $api.get_user_roles! }.to raise_error(Osm::NoActiveRoles)
+        expect($api).to receive(:post_query).with('api.php?action=getUserRoles').twice { fail Osm::OSMError, 'false' }
+        expect { $api.get_user_roles! }.to raise_error(Osm::OSMError::NoActiveRoles)
         expect($api.get_user_roles).to eq([])
       end
     end
 
     it 'Reraises any other Osm::Error' do
-      expect($api).to receive(:post_query).with('api.php?action=getUserRoles') { raise Osm::Error, 'Test' }
-      expect { $api.get_user_roles }.to raise_error(Osm::Error, 'Test')
+      expect($api).to receive(:post_query).with('api.php?action=getUserRoles') { raise Osm::OSMError, 'Test' }
+      expect { $api.get_user_roles }.to raise_error(Osm::OSMError, 'Test')
     end
 
   end