+++ spec/osm/flexi_record/column_spec.rb
@@ -119,7 +119,7 @@ describe Osm::FlexiRecord::Column do
         editable: false
       )
       expect($api).not_to receive(:post_query)
-      expect { col.update($api) }.to raise_error(Osm::Forbidden)
+      expect { col.update($api) }.to raise_error(Osm::OSMError::ReadOnly)
     end
 
   end # desxribe update
@@ -203,19 +203,9 @@ describe Osm::FlexiRecord::Column do
         editable: false
       )
       expect($api).not_to receive(:post_query)
-      expect { col.delete($api) }.to raise_error(Osm::Forbidden)
+      expect { col.delete($api) }.to raise_error(Osm::OSMError::ReadOnly)
     end
 
   end # desxribe delete
-
-
-
-
-
-
-
-
-
-
 
 end