+++ spec/osm/model_spec.rb
@@ -338,7 +338,7 @@ describe Osm::Model do
 
       it 'Raises exception when access is not allowed' do
         allow(Osm::Model).to receive('has_access_to_section?') { false }
-        expect { Osm::Model.require_access_to_section(api: $api, section: 5) }.to raise_error(Osm::Forbidden, 'You do not have access to that section')
+        expect { Osm::Model.require_access_to_section(api: $api, section: 5) }.to raise_error(Osm::OSMError::Forbidden, 'You do not have access to that section')
       end
 
     end
@@ -356,14 +356,14 @@ describe Osm::Model do
         allow(Osm::Model).to receive('user_has_permission?').and_return(false)
         allow(Osm::Model).to receive('api_has_permission?').and_return(true)
         section = Osm::Section.new(name: 'A SECTION')
-        expect { Osm::Model.require_permission(api: $api, to: :can_do, on: :can_on, section: section) }.to raise_error(Osm::Forbidden, 'Your OSM user does not have permission to can_do on can_on for A SECTION.')
+        expect { Osm::Model.require_permission(api: $api, to: :can_do, on: :can_on, section: section) }.to raise_error(Osm::OSMError::Forbidden, 'Your OSM user does not have permission to can_do on can_on for A SECTION.')
       end
 
       it "Raises exception when api doesn't have access" do
         allow(Osm::Model).to receive('user_has_permission?').and_return(true)
         allow(Osm::Model).to receive('api_has_permission?').and_return(false)
         section = Osm::Section.new(name: 'A SECTION')
-        expect { Osm::Model.require_permission(api: $api, to: :can_do, on: :can_on, section: section) }.to raise_error(Osm::Forbidden, 'You have not granted the can_do permissions on can_on to the API NAME API for A SECTION.')
+        expect { Osm::Model.require_permission(api: $api, to: :can_do, on: :can_on, section: section) }.to raise_error(Osm::OSMError::Forbidden, 'You have not granted the can_do permissions on can_on to the API NAME API for A SECTION.')
       end
 
     end
@@ -377,19 +377,19 @@ describe Osm::Model do
         section4 = Osm::Section.new(subscription_level: 4, name: 'NAME') # Gold+
 
         expect { Osm::Model.require_subscription(api: $api, level: 1, section: section1) }.not_to raise_error
-        expect { Osm::Model.require_subscription(api: $api, level: 2, section: section1) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Silver required for NAME).')
-        expect { Osm::Model.require_subscription(api: $api, level: 3, section: section1) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
-        expect { Osm::Model.require_subscription(api: $api, level: 4, section: section1) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: 2, section: section1) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Silver required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: 3, section: section1) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: 4, section: section1) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
 
         expect { Osm::Model.require_subscription(api: $api, level: 1, section: section2) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: 2, section: section2) }.not_to raise_error
-        expect { Osm::Model.require_subscription(api: $api, level: 3, section: section2) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
-        expect { Osm::Model.require_subscription(api: $api, level: 4, section: section2) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: 3, section: section2) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: 4, section: section2) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
 
         expect { Osm::Model.require_subscription(api: $api, level: 1, section: section3) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: 2, section: section3) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: 3, section: section3) }.not_to raise_error
-        expect { Osm::Model.require_subscription(api: $api, level: 4, section: section3) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: 4, section: section3) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
 
         expect { Osm::Model.require_subscription(api: $api, level: 1, section: section4) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: 2, section: section4) }.not_to raise_error
@@ -404,19 +404,19 @@ describe Osm::Model do
         section4 = Osm::Section.new(subscription_level: 4, name: 'NAME') # Gold+
 
         expect { Osm::Model.require_subscription(api: $api, level: :bronze, section: section1) }.not_to raise_error
-        expect { Osm::Model.require_subscription(api: $api, level: :silver, section: section1) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Silver required for NAME).')
-        expect { Osm::Model.require_subscription(api: $api, level: :gold, section: section1) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
-        expect { Osm::Model.require_subscription(api: $api, level: :gold_plus, section: section1) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: :silver, section: section1) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Silver required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: :gold, section: section1) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: :gold_plus, section: section1) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
 
         expect { Osm::Model.require_subscription(api: $api, level: :bronze, section: section2) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: :silver, section: section2) }.not_to raise_error
-        expect { Osm::Model.require_subscription(api: $api, level: :gold, section: section2) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
-        expect { Osm::Model.require_subscription(api: $api, level: :gold_plus, section: section2) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: :gold, section: section2) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: :gold_plus, section: section2) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
 
         expect { Osm::Model.require_subscription(api: $api, level: :bronze, section: section3) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: :silver, section: section3) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: :gold, section: section3) }.not_to raise_error
-        expect { Osm::Model.require_subscription(api: $api, level: :gold_plus, section: section3) }.to raise_error(Osm::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
+        expect { Osm::Model.require_subscription(api: $api, level: :gold_plus, section: section3) }.to raise_error(Osm::OSMError::Forbidden, 'Insufficent OSM subscription level (Gold+ required for NAME).')
 
         expect { Osm::Model.require_subscription(api: $api, level: :bronze, section: section4) }.not_to raise_error
         expect { Osm::Model.require_subscription(api: $api, level: :silver, section: section4) }.not_to raise_error